todo: finish hover impl
